# backend/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "BMSTU Web Course 2020 project",
    "title": "Forums API",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "paths": {
    "/api/v1/forum/:slug/create": {
      "post": {
        "description": "Добавление новой ветки обсуждения на форум.",
        "tags": [
          "forums"
        ],
        "summary": "Создание ветки",
        "operationId": "threadCreate",
        "parameters": [
          {
            "type": "string",
            "format": "identity",
            "description": "Идентификатор форума.",
            "name": "slug",
            "in": "path",
            "required": true
          },
          {
            "description": "Данные ветки обсуждения.",
            "name": "thread",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Thread"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Ветка обсуждения успешно создана. Возвращает данные созданной ветки обсуждения.",
            "schema": {
              "$ref": "#/definitions/Thread"
            }
          },
          "404": {
            "description": "Автор ветки или форум не найдены.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          },
          "409": {
            "description": "Ветка обсуждения уже существует. Возвращает данные ранее созданной ветки обсуждения.",
            "schema": {
              "$ref": "#/definitions/Thread"
            }
          }
        }
      }
    },
    "/api/v1/forum/:slug/details": {
      "get": {
        "description": "Получение информации о форуме по его идентификатору.",
        "tags": [
          "forums"
        ],
        "summary": "Получение информации о форуме",
        "operationId": "forumGetOne",
        "parameters": [
          {
            "type": "string",
            "format": "identity",
            "description": "Идентификатор форума.",
            "name": "slug",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о форуме.",
            "schema": {
              "$ref": "#/definitions/Forum"
            }
          },
          "404": {
            "description": "Форум отсутсвует в системе.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/api/v1/forum/:slug/threads": {
      "get": {
        "description": "Получение списка ветвей обсужления данного форума. Ветви обсуждения выводятся отсортированные по дате создания.",
        "tags": [
          "forums"
        ],
        "summary": "Список ветвей обсужления форума",
        "operationId": "forumGetThreads",
        "parameters": [
          {
            "type": "string",
            "description": "Идентификатор форума.",
            "name": "slug",
            "in": "path",
            "required": true
          },
          {
            "maximum": 10000,
            "minimum": 1,
            "type": "number",
            "default": 100,
            "description": "Максимальное кол-во возвращаемых записей.",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Дата создания ветви обсуждения, с которой будут выводиться записи (ветвь обсуждения с указанной датой попадает в результат выборки).",
            "name": "since",
            "in": "query"
          },
          {
            "type": "boolean",
            "description": "Флаг сортировки по убыванию.",
            "name": "desc",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о ветках обсуждения на форуме.",
            "schema": {
              "$ref": "#/definitions/Threads"
            }
          },
          "404": {
            "description": "Форум отсутсвует в системе.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/api/v1/forum/:slug/users": {
      "get": {
        "description": "Получение списка пользователей, у которых есть пост или ветка обсуждения в данном форуме. Пользователи выводятся отсортированные по nickname в порядке возрастания. Порядок сотрировки должен соответсвовать побайтовому сравнение в нижнем регистре.",
        "tags": [
          "forums"
        ],
        "summary": "Пользователи данного форума",
        "operationId": "forumGetUsers",
        "parameters": [
          {
            "type": "string",
            "format": "identity",
            "description": "Идентификатор форума.",
            "name": "slug",
            "in": "path",
            "required": true
          },
          {
            "maximum": 10000,
            "minimum": 1,
            "type": "number",
            "format": "int32",
            "default": 100,
            "description": "Максимальное кол-во возвращаемых записей.",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "string",
            "format": "identity",
            "description": "Идентификатор пользователя, с которого будут выводиться пользоватли (пользователь с данным идентификатором в результат не попадает).",
            "name": "since",
            "in": "query"
          },
          {
            "type": "boolean",
            "description": "Флаг сортировки по убыванию.",
            "name": "desc",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о пользователях форума.",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          },
          "404": {
            "description": "Форум отсутсвует в системе.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/api/v1/forum/create": {
      "post": {
        "description": "Создание нового форума.",
        "tags": [
          "forums"
        ],
        "summary": "Создание форума.",
        "operationId": "forumCreate",
        "parameters": [
          {
            "description": "Данные форума",
            "name": "forum",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ForumInput"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Форум успешно создан. Возвращает данные созданного форума.",
            "schema": {
              "$ref": "#/definitions/Forum"
            }
          },
          "400": {
            "description": "Владелец форума не найден",
            "schema": {
              "$ref": "#/responses/Error"
            }
          },
          "404": {
            "description": "Форум уже существует. Возвращает данные ранее созданного форума.",
            "schema": {
              "$ref": "#/definitions/Forum"
            }
          }
        }
      }
    },
...
    "/api/v1/post/:id/details": {
      "get": {
        "description": "Получение информации о ветке обсуждения по его имени.",
        "tags": [
          "posts"
        ],
        "summary": "Получение информации о ветке обсуждения",
        "operationId": "postGetOne",
        "parameters": [
          {
            "type": "number",
            "format": "int64",
            "description": "Идентификатор сообщения.",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "items": {
              "enum": [
                "user",
                "forum",
                "thread"
              ]
            },
            "description": "Включение полной информации о соответвующем объекте сообщения. Если тип объекта не указан, то полная информация об этих объектах не передаётся.",
            "name": "related",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о ветке обсуждения.",
            "schema": {
              "$ref": "#/definitions/PostFull"
            }
          },
          "404": {
            "description": "Ветка обсуждения отсутсвует в форуме.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      },
      "post": {
        "description": "Изменение сообщения на форуме. Если сообщение поменяло текст, то оно должно получить отметку `isEdited`.",
        "tags": [
          "posts"
        ],
        "summary": "Изменение сообщения",
        "operationId": "postUpdate",
        "parameters": [
          {
            "type": "number",
            "format": "int64",
            "description": "Идентификатор сообщения.",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Изменения сообщения.",
            "name": "post",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о сообщении.",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          },
          "404": {
            "description": "Сообщение отсутсвует в форуме.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
...
    "/api/v1/thread/:slug/create": {
      "post": {
        "description": "Добавление новых постов в ветку обсуждения на форум. Все посты, созданные в рамках одного вызова данного метода должны иметь одинаковую дату создания (Post.Created).",
        "tags": [
          "threads"
        ],
        "summary": "Создание новых постов",
        "operationId": "postsCreate",
        "parameters": [
          {
            "type": "string",
            "format": "identity",
            "description": "Идентификатор ветки обсуждения.",
            "name": "slug_or_id",
            "in": "path",
            "required": true
          },
          {
            "description": "Список создаваемых постов.",
            "name": "posts",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Posts"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Посты успешно созданы. Возвращает данные созданных постов в том же порядке, в котором их передали на вход метода.",
            "schema": {
              "$ref": "#/definitions/Posts"
            }
          },
          "404": {
            "description": "Ветка обсуждения отсутствует в базе данных.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          },
          "409": {
            "description": "Хотя бы один родительский пост отсутсвует в текущей ветке обсуждения.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/api/v1/thread/:slug/details": {
      "get": {
        "description": "Получение информации о ветке обсуждения по его имени.",
        "tags": [
          "threads"
        ],
        "summary": "Получение информации о ветке обсуждения",
        "operationId": "threadGetOne",
        "parameters": [
          {
            "type": "string",
            "description": "Идентификатор ветки обсуждения.",
            "name": "slug_or_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о ветке обсуждения.",
            "schema": {
              "$ref": "#/definitions/Thread"
            }
          },
          "404": {
            "description": "Ветка обсуждения отсутсвует в форуме.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      },
      "post": {
        "description": "Обновление ветки обсуждения на форуме.",
        "tags": [
          "threads"
        ],
        "summary": "Обновление ветки",
        "operationId": "threadUpdate",
        "parameters": [
          {
            "type": "string",
            "format": "identity",
            "description": "Идентификатор ветки обсуждения.",
            "name": "slug_or_id",
            "in": "path",
            "required": true
          },
          {
            "description": "Данные ветки обсуждения.",
            "name": "thread",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ThreadUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о ветке обсуждения.",
            "schema": {
              "$ref": "#/definitions/Thread"
            }
          },
          "404": {
            "description": "Ветка обсуждения отсутсвует в форуме.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/api/v1/thread/:slug/posts": {
      "get": {
        "description": "Получение списка сообщений в данной ветке форуме. Сообщения выводятся отсортированные по дате создания.",
        "tags": [
          "threads"
        ],
        "summary": "Сообщения данной ветви обсуждения",
        "operationId": "threadGetPosts",
        "parameters": [
          {
            "type": "string",
            "format": "identity",
            "description": "Идентификатор ветки обсуждения.",
            "name": "slug_or_id",
            "in": "path",
            "required": true
          },
          {
            "maximum": 10000,
            "minimum": 1,
            "type": "number",
            "format": "int32",
            "default": 100,
            "description": "Максимальное кол-во возвращаемых записей.",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "number",
            "format": "int64",
            "description": "Идентификатор поста, после которого будут выводиться записи (пост с данным идентификатором в результат не попадает).",
            "name": "since",
            "in": "query"
          },
          {
            "enum": [
              "flat",
              "tree",
              "parent_tree"
            ],
            "type": "string",
            "default": "flat",
            "description": "Вид сортировки:\n* flat - по дате, комментарии выводятся простым списком в порядке создания;\n* tree - древовидный, комментарии выводятся отсортированные в дереве по N штук; \n* parent_tree - древовидные с пагинацией по родительским (parent_tree), на странице N родительских комментов и все комментарии прикрепленные к ним, в древвидном отображение.",
            "name": "sort",
            "in": "query"
          },
          {
            "type": "boolean",
            "description": "Флаг сортировки по убыванию.",
            "name": "desc",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о сообщениях форума.",
            "schema": {
              "$ref": "#/definitions/Posts"
            }
          },
          "404": {
            "description": "Ветка обсуждения отсутсвует в форуме.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/api/v1/thread/:slug/vote": {
      "post": {
        "description": "Изменение голоса за ветвь обсуждения. Один пользователь учитывается только один раз и может изменить своё мнение.",
        "tags": [
          "threads"
        ],
        "summary": "Проголосовать за ветвь обсуждения",
        "operationId": "threadVote",
        "parameters": [
          {
            "type": "string",
            "format": "identity",
            "description": "Идентификатор ветки обсуждения.",
            "name": "slug_or_id",
            "in": "path",
            "required": true
          },
          {
            "description": "Информация о голосовании пользователя.",
            "name": "vote",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Vote"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о ветке обсуждения.",
            "schema": {
              "$ref": "#/definitions/Thread"
            }
          },
          "404": {
            "description": "Ветка обсуждения отсутсвует в форуме.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "PostUpdate": {
      "description": "Сообщение для обновления существующего",
      "type": "object",
      "properties": {
        "message": {
          "description": "Собственно сообщение",
          "type": "string",
          "x-go-name": "Message",
          "example": "We should be afraid of the Kraken"
        }
      },
      "x-go-name": "postInput",
      "x-go-package": "dbProject/forum/delivery/http"
    },
...
    "Thread": {
      "description": "Ветка обсуждений на форуме",
      "type": "object",
      "properties": {
        "author": {
          "description": "Пользователь, создавший ветку\n\nformat: identity",
          "type": "string",
          "x-go-name": "Author",
          "example": "j.sparrow"
        },
        "created": {
          "description": "Дата создания ветки",
          "type": "string",
          "format": "date-time",
          "x-go-name": "Created",
          "example": "2017-01-01T00:00:00.000Z"
        },
        "forum": {
          "description": "Форум ветки\n\nformat: identity",
          "type": "string",
          "x-go-name": "ForumName",
          "readOnly": true,
          "example": "pirate-stories"
        },
        "id": {
          "description": "Идентификатор ветки обсуждения",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Id",
          "readOnly": true,
          "example": 42
        },
        "message": {
          "description": "Описание ветки обсуждения",
          "type": "string",
          "x-go-name": "Message",
          "example": "An urgent need to reveal the hiding place of Davy Jones. Who is willing to help in this matter?"
        },
        "posts": {
          "description": "Количество сообщений в ветке",
          "type": "integer",
          "format": "int64",
          "x-go-name": "PostsCount",
          "readOnly": true,
          "example": 100
        },
        "slug": {
          "description": "Человекопонятный URL\n\nformat: identity",
          "type": "string",
          "x-go-name": "Slug",
          "readOnly": true,
          "example": "jones-cache"
        },
        "title": {
          "description": "Заголовок ветки",
          "type": "string",
          "x-go-name": "Title",
          "example": "Davy Jones cache"
        },
        "votes": {
          "description": "Кол-во голосов",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Votes",
          "readOnly": true
        }
      },
      "x-go-package": "dbProject/forum/delivery/http"
    },
...
    "Threads": {
      "type": "object",
      "properties": {
        "threads": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Post"
          },
          "x-go-name": "Threads"
        }
      },
      "x-go-package": "dbProject/forum/delivery/http"
    },
...
    "Users": {
      "type": "object",
      "properties": {
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserOutput"
          },
          "x-go-name": "Users"
        }
      },
      "x-go-package": "dbProject/user/delivery/http"
    },
...
  },
  "responses": {
    "Error": {
      "description": "Error response payload",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "description": "Текстовое описание ошибки.",
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "KEY",
      "in": "header"
    }
  },
  "security": [
    {
      "api_key": []
    }
  ]
}